Change - mock::config_opts replaces mock::site_defaults
The `site_defaults` parameter specifying a source URI to a replete
site-defaults.cfg file has replaced by the new `config_opts` parameter
that as a hash-map is intended to provide every value that is to
override the package defaults for this file.  The present defaults match
what mock-1.4.7-1.fc27.noarch and mock-1.4.7-2.el7.noarch ship for
Fedora 27 and CentOS 7.4 respectively.  Thus, if nothing was specified
for the old parameter, specifying nothing for the new one will result in
the same configuration, although the file will look much differently due
to the eliminate of so many comments.

Signed-off-by: John Florian <jflorian@doubledog.org>