Bug - /etc/mock/default.cfg shouldn't be a link
The standard install of mock (or more recently mock-core-configs) has
this as a symlink targeting the config matching the host platform.
That's convenient for the average user, but if Puppet is being used to
manage this, things can end up being a bit deceptive.  To resolve this,
the link will now be replaced with a regular file and the expected
content w/o altering some other config unintentionally.

Signed-off-by: John Florian <jflorian@doubledog.org>